Move logging toggle to a safer location
