+++ source/index.ts
@@ -4,7 +4,7 @@ export * from "./receiver.js";
 export * from "./sender.js";
 export * from "./types.js";
 export { getThisFrame, getTopLevelFrame } from "./thisTarget.js";
-export { toggleLogging } from "./shared.js";
+export { toggleLogging } from "./logging.js";
 
 import { initPrivateApi } from "./thisTarget.js";
