+++ source/shared.ts
@@ -13,30 +13,12 @@ export function isObject(value: unknown): value is Record<string, unknown> {
   return typeof value === "object" && value !== null;
 }
 
-function noop() {
-  /* */
-}
-
 export class MessengerError extends Error {
   override name = "MessengerError";
 }
 
 // @ts-expect-error Wrong `errorConstructors` types
 errorConstructors.set("MessengerError", MessengerError);
-
-// .bind preserves the call location in the console
-const debug = console.debug.bind(console, "Messenger:");
-const warn = console.warn.bind(console, "Messenger:");
-
-export const log = { debug, warn };
-
-export function toggleLogging(enabled: boolean): void {
-  log.debug = enabled ? debug : noop;
-  log.warn = enabled ? warn : noop;
-}
-
-// Default to "no logs"
-toggleLogging(false);
 
 export function isErrorObject(error: unknown): error is ErrorObject {
   // eslint-disable-next-line @typescript-eslint/no-explicit-any -- This is a type guard function and it uses ?.
